+++ core/src/test/java/arez/spy/ObserveScheduledEventTest.java
@@ -9,7 +9,7 @@ import org.realityforge.guiceyloops.shared.ValueUtil;
 import org.testng.annotations.Test;
 import static org.testng.Assert.*;
 
-public class ReactionScheduledEventTest
+public class ObserveScheduledEventTest
   extends AbstractArezTest
 {
   @Test
@@ -19,14 +19,14 @@ public class ReactionScheduledEventTest
     final String name = "Foo@1";
     final Observer observer = context.tracker( name, ValueUtil::randomString );
     final ObserverInfo info = context.getSpy().asObserverInfo( observer );
-    final ReactionScheduledEvent event = new ReactionScheduledEvent( info );
+    final ObserveScheduledEvent event = new ObserveScheduledEvent( info );
 
     assertEquals( event.getObserver(), info );
 
     final HashMap<String, Object> data = new HashMap<>();
     event.toMap( data );
 
-    assertEquals( data.get( "type" ), "ReactionScheduled" );
+    assertEquals( data.get( "type" ), "ObserveScheduled" );
     assertEquals( data.get( "name" ), name );
     assertEquals( data.size(), 2 );
   }